backend/requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
pytz==2018.5
uWSGI==2.0.15